cleanup ultrasonic distance sensor
The pulse duration should ideally be 10us. It is currently set to 1ms as
this is the shortest delay. (TODO)
Read back values are not accurate. Fixable by multiplying us with 1.5.
If no us is read back or it is too large do not return "0" but return
the maximum value. Altough return 0 is noce for "c" developpers. It is
not a good choice or makecode